fix(writer): encode delimiters as UTF-8
